requirements.txt
absl-py==2.1.0
altair==5.3.0
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.14.0
flatbuffers==24.3.25
fonttools==4.50.0
fsspec==2024.5.0
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
google_images_download==2.8.0
grpcio==1.64.0
h11==0.14.0
h5py==3.11.0
idna==3.6
import-ipynb==0.1.4
intel-openmp==2021.4.0
ipykernel==6.29.4
ipython==8.23.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.3.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
keras==3.3.3
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.6
mdurl==0.1.2
mkl==2021.4.0
ml-dtypes==0.3.2
mpmath==1.3.0
namex==0.0.8
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.1
parso==0.8.4
pillow==10.3.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
pyarrow==16.1.0
pycparser==2.21
pydeck==0.9.1
Pygments==2.17.2
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
pyzmq==25.1.2
referencing==0.35.1
requests==2.32.2
rich==13.7.1
rpds-py==0.18.1
scikit-learn==1.4.1.post1
scipy==1.13.0
seaborn==0.13.2
selenium==4.18.1
setuptools==70.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
streamlit==1.35.0
sympy==1.12.1
tbb==2021.12.0
tenacity==8.3.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
termcolor==2.4.0
threadpoolctl==3.4.0
timeout-decorator==0.5.0
toml==0.10.2
toolz==0.12.1
torch==2.3.0
torchvision==0.18.0
tornado==6.4
traitlets==5.14.2
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
watchdog==4.0.1
wcwidth==0.2.13
webdriver-manager==4.0.1
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
wsproto==1.2.0
